test: 도메인 유효성 검증 테스트 추가
